requirements.txt
Flask==2.3.2
pyotp==2.8.0
qrcode[pil]==7.4.2
python-dotenv==1.0.0
cryptography==41.0.3
